+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 
 # Changelog
 
+[1.6.2+hotfix.1] - 2026-09-08
+
+* [Windows/Linux] fix(crash): Do not access the WebRTC API before calling EnsureWebRTCInitialized (#2169)
+
+[1.6.2] - 2026-09-07
+
+* [Darwin/Android/Windows/Linux] Add WARP support.
+
 [1.6.1] - 2026-09-01
 
 * [Darwin/Android/Windows/Linux] feat: upgrade libwebrtc to 150.7871.01.